Description not generated from from the product title, refactoring JS code
