--- plugins/woo-ai/src/components/index.ts
@@ -1,2 +1,0 @@
-export * from './woo-ai-button-item';
-export * from './woo-ai-header-item-fills';
