--- plugins/woo-ai/src/components/woo-ai-button-item/index.ts
@@ -1,1 +1,0 @@
-export * from './woo-ai-button-item';
