--- plugins/woo-ai/src/components/woo-ai-button-item/woo-ai-button-item.tsx
@@ -1,34 +1,0 @@
-/**
- * External dependencies
- */
-import {
-	createSlotFill,
-	Slot as BaseSlot,
-	Fill as BaseFill,
-} from '@wordpress/components';
-
-type WooAIButtonItemSlot = React.FC< BaseSlot.Props >;
-
-type WooAIButtonItemFill = React.FC< BaseFill.Props > & {
-	Slot?: WooAIButtonItemSlot;
-};
-
-type CreateSlotFillReturn = {
-	Fill: WooAIButtonItemFill;
-	Slot: WooAIButtonItemSlot;
-};
-
-const { Fill, Slot }: CreateSlotFillReturn =
-	createSlotFill( 'WooAIButtonItem' );
-
-Fill.Slot = ( { fillProps } ) => (
-	<Slot fillProps={ fillProps }>
-		{ ( fills ) => {
-			return fills.length === 0 ? null : <>{ fills }</>;
-		} }
-	</Slot>
-);
-
-export const WooAIButtonItem = Fill as WooAIButtonItemFill & {
-	Slot: WooAIButtonItemSlot;
-};
