--- plugins/woo-ai/src/components/woo-ai-header-item-fills/index.ts
@@ -1,1 +1,0 @@
-export * from './woo-ai-header-item-fills';
