--- plugins/woo-ai/src/components/woo-ai-header-item-fills/woo-ai-header-item-fills.tsx
@@ -1,27 +1,0 @@
-/**
- * External dependencies
- */
-import { registerPlugin } from '@wordpress/plugins';
-import { __ } from '@wordpress/i18n';
-
-/**
- * Internal dependencies
- */
-import { WooAIButtonItem } from '../index';
-
-registerPlugin( 'woocommerce-ai-product-description-header', {
-	// @ts-expect-error 'scope' does exist. @types/wordpress__plugins is outdated.
-	scope: 'woocommerce-ai-product-description',
-	render: () => (
-		<>
-			<WooAIButtonItem>
-				<button
-					className="button wp-media-button wc-write-it-for-me"
-					type="button"
-				>
-					{ __( 'Write it for me (beta)', 'woocommerce' ) }
-				</button>
-			</WooAIButtonItem>
-		</>
-	),
-} );
