+++ plugins/woo-ai/src/product-description/index.ts
@@ -1,2 +1,1 @@
 export * from './product-description-button-container';
-export * from './product-description-form';
